% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigeon_cleanr.R
\name{pigeon_clean}
\alias{pigeon_clean}
\title{Cleans raw data to be ready for processing}
\usage{
pigeon_clean(x, method = "default")
}
\arguments{
\item{x}{List of raw data}

\item{method}{Type of raw data being clean}
}
\value{
A dataframe with usable, organised, and standardized but not processed data.
}
\description{
Cleans raw data to be ready for processing
}
