% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigeon_quickr.R
\name{pigeon_quick}
\alias{pigeon_quick}
\title{Quickly takes raw data to processed data}
\usage{
pigeon_quick(method, patterns.regex = c(rep(NULL, length(method))),
  patterns.exc = c(rep(NULL, length(method))), path = c(rep(getwd(),
  length(method))))
}
\arguments{
\item{method}{vector of types of data being used}

\item{patterns.regex}{vector of Regular Expression to find files}

\item{patterns.exc}{vector of Regular Expression to exclude from files}

\item{path}{vector of file locations}
}
\value{
A dataframe with aggregated looking, widened data, and trial information
}
\description{
Quickly takes raw data to processed data
}
